split some skip_xxx functions from combine.cpp
