+++ src/combine.cpp
@@ -12,6 +12,7 @@
 
 #include "chunk_list.h"
 #include "combine_labels.h"
+#include "combine_skip.h"
 #include "ChunkStack.h"
 #include "error_types.h"
 #include "flag_parens.h"
@@ -46,26 +47,6 @@ static void flag_asm(chunk_t *pc);
 
 //! Scan backwards to see if we might be on a type declaration
 static bool chunk_ends_type(chunk_t *start);
-
-
-//! Skips to the start of the next statement.
-static chunk_t *skip_to_next_statement(chunk_t *pc);
-
-
-/**
- * Skips everything until a comma or semicolon at the same level.
- * Returns the semicolon, comma, or close brace/paren or nullptr.
- */
-static chunk_t *skip_expression(chunk_t *start);
-
-
-/**
- * Skips the D 'align()' statement and the colon, if present.
- *    align(2) int foo;  -- returns 'int'
- *    align(4):          -- returns 'int'
- *    int bar;
- */
-static chunk_t *skip_align(chunk_t *start);
 
 
 /**
@@ -3895,19 +3876,6 @@ static void fix_fcn_def_params(chunk_t *start)
 } // fix_fcn_def_params
 
 
-static chunk_t *skip_to_next_statement(chunk_t *pc)
-{
-   while (  pc != nullptr
-         && !chunk_is_semicolon(pc)
-         && pc->type != CT_BRACE_OPEN
-         && pc->type != CT_BRACE_CLOSE)
-   {
-      pc = chunk_get_next_ncnl(pc);
-   }
-   return(pc);
-}
-
-
 static chunk_t *fix_variable_definition(chunk_t *start)
 {
    LOG_FUNC_ENTRY();
@@ -4067,23 +4035,6 @@ static chunk_t *fix_variable_definition(chunk_t *start)
    }
    return(skip_to_next_statement(end));
 } // fix_variable_definition
-
-
-static chunk_t *skip_expression(chunk_t *start)
-{
-   chunk_t *pc = start;
-
-   while (pc != nullptr && pc->level >= start->level)
-   {
-      if (  pc->level == start->level
-         && (chunk_is_semicolon(pc) || chunk_is_token(pc, CT_COMMA)))
-      {
-         return(pc);
-      }
-      pc = chunk_get_next_ncnl(pc);
-   }
-   return(pc);
-}
 
 
 bool go_on(chunk_t *pc, chunk_t *start)
@@ -5659,76 +5610,6 @@ static void mark_class_ctor(chunk_t *start)
 } // mark_class_ctor
 
 
-static chunk_t *skip_align(chunk_t *start)
-{
-   chunk_t *pc = start;
-
-   if (chunk_is_token(pc, CT_ALIGN))
-   {
-      pc = chunk_get_next_ncnl(pc);
-
-      if (chunk_is_token(pc, CT_PAREN_OPEN))
-      {
-         pc = chunk_get_next_type(pc, CT_PAREN_CLOSE, pc->level);
-         pc = chunk_get_next_ncnl(pc);
-
-         if (chunk_is_token(pc, CT_COLON))
-         {
-            pc = chunk_get_next_ncnl(pc);
-         }
-      }
-   }
-   return(pc);
-}
-
-
-chunk_t *skip_parent_types(chunk_t *colon)
-{
-   auto pc = chunk_get_next_ncnlnp(colon);
-
-   while (pc)
-   {
-      // Skip access specifier
-      if (chunk_is_token(pc, CT_ACCESS))
-      {
-         pc = chunk_get_next_ncnlnp(pc);
-         continue;
-      }
-
-      // Check for a type name
-      if (!(chunk_is_token(pc, CT_WORD) || chunk_is_token(pc, CT_TYPE)))
-      {
-         LOG_FMT(LPCU,
-                 "%s is confused; expected a word at %zu:%zu "
-                 "following type list at %zu:%zu\n", __func__,
-                 colon->orig_line, colon->orig_col,
-                 pc->orig_line, pc->orig_col);
-         return(colon);
-      }
-      // Get next token
-      auto next = skip_template_next(chunk_get_next_ncnlnp(pc));
-
-      if (chunk_is_token(next, CT_DC_MEMBER) || chunk_is_token(next, CT_COMMA))
-      {
-         pc = chunk_get_next_ncnlnp(next);
-      }
-      else if (next)
-      {
-         LOG_FMT(LPCU, "%s -> %zu:%zu ('%s')\n", __func__,
-                 next->orig_line, next->orig_col, next->text());
-         return(next);
-      }
-      else
-      {
-         break;
-      }
-   }
-   LOG_FMT(LPCU, "%s: did not find end of type list (start was %zu:%zu)\n",
-           __func__, colon->orig_line, colon->orig_col);
-   return(colon);
-} // skip_parent_types
-
-
 static void mark_struct_union_body(chunk_t *start)
 {
    LOG_FUNC_ENTRY();
@@ -6280,66 +6161,6 @@ chunk_t *skip_template_next(chunk_t *ang_open)
       return(chunk_get_next_ncnl(pc));
    }
    return(ang_open);
-}
-
-
-chunk_t *skip_template_prev(chunk_t *ang_close)
-{
-   if (chunk_is_token(ang_close, CT_ANGLE_CLOSE))
-   {
-      chunk_t *pc = chunk_get_prev_type(ang_close, CT_ANGLE_OPEN, ang_close->level);
-      return(chunk_get_prev_ncnlni(pc));   // Issue #2279
-   }
-   return(ang_close);
-}
-
-
-chunk_t *skip_tsquare_next(chunk_t *ary_def)
-{
-   if (chunk_is_token(ary_def, CT_SQUARE_OPEN) || chunk_is_token(ary_def, CT_TSQUARE))
-   {
-      return(chunk_get_next_nisq(ary_def));
-   }
-   return(ary_def);
-}
-
-
-chunk_t *skip_attribute_next(chunk_t *attr)
-{
-   chunk_t *pc = attr;
-
-   while (chunk_is_token(pc, CT_ATTRIBUTE))
-   {
-      pc = chunk_get_next_ncnl(pc);
-
-      if (chunk_is_token(pc, CT_FPAREN_OPEN))
-      {
-         pc = chunk_get_next_type(pc, CT_FPAREN_CLOSE, pc->level);
-         pc = chunk_get_next_ncnl(pc);
-      }
-   }
-   return(pc);
-}
-
-
-chunk_t *skip_attribute_prev(chunk_t *fp_close)
-{
-   chunk_t *pc = fp_close;
-
-   while (true)
-   {
-      if (  chunk_is_token(pc, CT_FPAREN_CLOSE)
-         && get_chunk_parent_type(pc) == CT_ATTRIBUTE)
-      {
-         pc = chunk_get_prev_type(pc, CT_ATTRIBUTE, pc->level);
-      }
-      else if (chunk_is_not_token(pc, CT_ATTRIBUTE))
-      {
-         break;
-      }
-      pc = chunk_get_prev_ncnlni(pc);   // Issue #2279
-   }
-   return(pc);
 }
