+++ src/combine.h
@@ -77,43 +77,4 @@ chunk_t *set_paren_parent(chunk_t *start, c_token_t parent);
 void do_symbol_check(chunk_t *prev, chunk_t *pc, chunk_t *next);
 
 
-/**
- * Skips over the rest of the template if ang_open is indeed a CT_ANGLE_OPEN.
- * Points to the chunk after the CT_ANGLE_CLOSE.
- * If the chunk isn't an CT_ANGLE_OPEN, then it is returned.
- */
-chunk_t *skip_template_next(chunk_t *ang_open);
-
-
-/**
- * Skips over the rest of the template if ang_close is indeed a CT_ANGLE_CLOSE.
- * Points to the chunk before the CT_ANGLE_OPEN
- * If the chunk isn't an CT_ANGLE_CLOSE, then it is returned.
- */
-chunk_t *skip_template_prev(chunk_t *ang_close);
-
-
-/**
- * Skips the rest of the array definitions if ary_def is indeed a
- * CT_TSQUARE or CT_SQUARE_OPEN
- */
-chunk_t *skip_tsquare_next(chunk_t *ary_def);
-
-
-/**
- * If attr is CT_ATTRIBUTE, then skip it and the parens and return the chunk
- * after the CT_FPAREN_CLOSE.
- * If the chunk isn't an CT_ATTRIBUTE, then it is returned.
- */
-chunk_t *skip_attribute_next(chunk_t *attr);
-
-
-/**
- * If fp_close is a CT_FPAREN_CLOSE with a parent of CT_ATTRIBUTE, then skip it
- * and the '__attribute__' thingy and return the chunk before CT_ATTRIBUTE.
- * Otherwise return fp_close.
- */
-chunk_t *skip_attribute_prev(chunk_t *fp_close);
-
-
 #endif /* COMBINE_H_INCLUDED */
